refactor: extract templates and separate logic in lp-generator
