+++ package.json
@@ -12,6 +12,7 @@
     "format": "prettier . --check --experimental-cli",
     "format:fix": "prettier . --write --experimental-cli",
     "lint": "biome lint .",
+    "test": "vitest",
     "validate": "run-p format lint typecheck"
   },
   "dependencies": {
@@ -79,7 +80,8 @@
     "typescript": "^5.9.3",
     "vite": "^7.2.2",
     "vite-plugin-devtools-json": "1.0.0",
-    "vite-tsconfig-paths": "^5.1.4"
+    "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "4.0.10"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
